backend/requirements.txt
firebase-admin>=6.0.0
google-cloud-firestore>=2.1.0
google-cloud-language>=2.0.0
Flask>=2.0.0
